requirements.txt
BeautifulSoup==3.2.1
cffi==1.1.2
cryptography==0.9.3
Django==1.8.4
django-appconf==1.0.1
django-boss==0.6.3
django-bower==5.0.4
django-cache-utils==0.7.2
django-jsonfield==0.9.15
django-likes==0.2
django-mailer==1.1
django-mediagenerator==1.11
django-mptt==0.7.4
django-mptt-admin==0.2.1
django-mptt-urls==1.0
django-nocaptcha-recaptcha==0.0.18
django-nvd3==0.8.2
django-qsstats-magic==0.7.2
django-simple-captcha==0.4.6
django-taggit==0.17.1
django-templatetag-sugar==1.0
django-widget-tweaks==1.4.1
enum34==1.0.4
google-api-python-client==1.4.2
gunicorn==19.3.0
httplib2==0.9.2
idna==2.0
ipaddress==1.0.11
Jinja2==2.8
lockfile==0.10.2
MarkupSafe==0.23
MySQL-python==1.2.5
ndg-httpsclient==0.4.0
oauth2client==1.5.1
Pillow==2.9.0
psycopg2==2.6.1
pyasn1==0.1.8
pyasn1-modules==0.0.8
pycparser==2.14
pyOpenSSL==0.15.1
python-dateutil==2.4.2
python-memcached==1.57
python-nvd3==0.13.7
python-slugify==0.0.9
pytz==2015.4
rsa==3.2
setproctitle==1.1.9
simplejson==3.8.0
six==1.9.0
sorl-thumbnail==12.3
Unidecode==0.4.18
uritemplate==0.6
versiontools==1.9.1
wheel==0.24.0
